+++ src/users/index.ts
@@ -1,39 +1,94 @@
-import * as tools from 'firebase-tools'
 import * as admin from 'firebase-admin'
-import { tmpdir } from 'os'
-import { resolve, parse } from 'path'
-import operationResponse, { UsersStatusResponse } from '../_lib/operation'
-import asyncMiddleware from '../_lib/asyncMiddleware'
-import { promisify } from 'util'
+import * as tools from 'firebase-tools'
 import fs from 'fs'
+import fetch from 'node-fetch'
+import { tmpdir } from 'os'
+import { parse, resolve } from 'path'
+import { format } from 'url'
+import { promisify } from 'util'
+import { defaultControllerDomain } from '../options'
+import asyncMiddleware from '../_lib/asyncMiddleware'
+import operationResponse, { UsersStatusResponse } from '../_lib/operation'
 
 const unlink = promisify(fs.unlink)
 
-export type UsersBackupOptions = {
+export interface UsersBackupOptions {
   bucketsAllowlist?: string[]
   projectId: string
+  controllerDomain?: string
+  controllerToken: string
+  agentURL: string
 }
 
-export type UsersBackupRequestOptions = {
+export interface UsersBackupRequestBody {
   storageId: string
   path: string
+  delay?: {
+    backupId: string
+    state: 'delay' | 'backup'
+  }
 }
 
 export function backupUsersMiddleware({
   bucketsAllowlist,
-  projectId
+  projectId,
+  controllerDomain,
+  controllerToken,
+  agentURL,
 }: UsersBackupOptions) {
   return asyncMiddleware(async (request, response) => {
     // TODO: Validate options
-    const options = request.body as UsersBackupRequestOptions
-    const state = await backupUsers(projectId, options)
-    operationResponse(response, state)
+    const body = request.body as UsersBackupRequestBody
+
+    if (body.delay?.state === 'delay') {
+      // NOTE: Trigger backup, but do not wait for the result
+      fetch(agentURL + request.path, {
+        method: 'POST',
+        headers: {
+          // Authorization can't be missing as we verify it
+          Authorization: request.header('Authorization')!,
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          storageId: body.storageId,
+          path: body.path,
+          delay: {
+            state: 'backup',
+            backupId: body.delay.backupId,
+          },
+        }),
+      })
+
+      operationResponse(response, { state: 'pending' })
+    } else {
+      const backupResponse = await backupUsers(projectId, body)
+
+      if (body.delay) {
+        const reportURL = format({
+          hostname: controllerDomain || defaultControllerDomain,
+          protocol: 'https',
+          pathname: '/reportBackup',
+        })
+        await fetch(reportURL, {
+          method: 'POST',
+          headers: { 'Content-Type': 'application/json' },
+          body: JSON.stringify({
+            token: controllerToken,
+            backupId: body.delay.backupId,
+            type: 'users',
+            ...backupResponse,
+          }),
+        })
+      }
+
+      operationResponse(response, backupResponse)
+    }
   })
 }
 
 async function backupUsers(
   projectId: string,
-  options: UsersBackupRequestOptions
+  options: UsersBackupRequestBody
 ): Promise<UsersStatusResponse> {
   // Create bucket
   const bucket = admin.storage().bucket(options.storageId)
@@ -51,7 +106,7 @@ async function backupUsers(
     calculateUsers(path),
     bucket
       .upload(path, { destination: options.path })
-      .then(([file]) => file.metadata.size as string)
+      .then(([file]) => file.metadata.size as string),
   ])
 
   // Remove the temporary file
@@ -66,12 +121,25 @@ async function backupUsers(
  * @returns the number of users in the backup
  */
 export async function calculateUsers(path: string) {
-  const usersStream = fs.createReadStream(path, {
-    encoding: 'utf8',
-    highWaterMark: 10000000 // 10MB
-  })
+  let usersCount = 0
+  let lookingForEnding: string | null = null
 
-  return calculateUsersInSteam(usersStream)
+  for await (const chunk of generateFileChunks(path, 10000000 /* 10MB */)) {
+    const text: string = chunk.toString()
+
+    if (lookingForEnding) {
+      if (text.slice(0, lookingForEnding.length) === lookingForEnding)
+        usersCount++
+      lookingForEnding = null
+    }
+
+    usersCount += text.match(/"localId"/g)?.length || 0
+
+    const ending = text.match(/"(l(o(c(a(l(I(d(")?)?)?)?)?)?)?)?$/)
+    if (ending) lookingForEnding = '"localId"'.slice(ending[0].length)
+  }
+
+  return usersCount
 }
 
 /**
@@ -99,6 +167,52 @@ export async function calculateUsersInSteam(usersStream: fs.ReadStream) {
   }
 
   return usersCount
+}
+
+/**
+ * Genenerates chunks of data from the given file.
+ *
+ * The code is based on the article by Kasper Moskwiak (https://github.com/kmoskwiak): https://betterprogramming.pub/a-memory-friendly-way-of-reading-files-in-node-js-a45ad0cc7bb6
+ *
+ * @param path - the file path to read
+ * @param size - the chunk size
+ */
+async function* generateFileChunks(path: string, size: number) {
+  const sharedBuffer = Buffer.alloc(size)
+  const stats = fs.statSync(path)
+  const file = fs.openSync(path, 'r')
+
+  let bytesRead = 0 // How many bytes were read
+  let end = size
+
+  for (let chunk = 0; chunk < Math.ceil(stats.size / size); chunk++) {
+    await readFileBytes(file, sharedBuffer)
+
+    bytesRead = (chunk + 1) * size
+    // When we reach the end of file, we have to calculate how many bytes were
+    // actually read.
+    if (bytesRead > stats.size) end = size - (bytesRead - stats.size)
+
+    yield sharedBuffer.slice(0, end)
+  }
+}
+
+/**
+ * Reads the file bytes into the shared buffer.
+ *
+ * The code is based on the article by Kasper Moskwiak (https://github.com/kmoskwiak): https://betterprogramming.pub/a-memory-friendly-way-of-reading-files-in-node-js-a45ad0cc7bb6
+ *
+ * @param file - the file descriptor
+ * @param buffer - the shared buffer to use
+ * @returns promise to file read
+ */
+function readFileBytes(file: number, buffer: Buffer) {
+  return new Promise((resolve, reject) => {
+    fs.read(file, buffer, 0, buffer.length, null, (error) => {
+      if (error) return reject(error)
+      resolve(void 0)
+    })
+  })
 }
 
 /**